+++ apk/Changelog.md
@@ -9,6 +9,16 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Web Version July 12 2022
+==============
+
+* Update User channel screen to use new Twitch API, the sorting now is base on the last channel you followed to show first, on the future this list may get a better sorting today the APIs don't support it
+* Update Search Game to use new Twitch API
+* Update Search Channel to use new Twitch API
+* Remove Search Live no longer supported by new Twitch API
+* Update Feature screen to use new Twitch API
+* General improves
+
 Web Version July 11 2022
 ==============